Fix color in app navbar search
