+++ knowledge/src/brandingTheme.ts
@@ -1,4 +1,3 @@
-import { grey } from "@mui/material/colors";
 import { Theme, ThemeOptions } from "@mui/material/styles";
 
 declare module "@mui/material/styles/createPalette" {
@@ -21,6 +20,19 @@ const common = {
   orangeDark: "#ff6d00",
   darkGrayBackground: "#28282A",
   gray: "#D3D3D3"
+};
+
+const grey = {
+  50: "#fafafa",
+  100: "#f8f8f8",
+  200: "#eeeeee",
+  300: "#e0e0e0",
+  400: "#bdbdbd",
+  500: "#9e9e9e",
+  600: "#757575",
+  700: "#616161",
+  800: "#424242",
+  900: "#212121"
 };
 
 const systemFont = ["Roboto", "sans-serif"];
@@ -70,7 +82,8 @@ export const getDesignTokens = (mode: "light" | "dark") =>
           primary: common.white,
           secondary: grey[300]
         }
-      })
+      }),
+      grey
     },
     typography: {
       fontFamily: [...systemFont].join(","),
